requirements.txt
bidict==0.21.2
click==7.1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
timeloop==1.0.2
Werkzeug==1.0.1
